Finished a usable version of the bot

+++ Config.cs
@@ -3,7 +3,7 @@
     internal static class Config
     {
         internal static string? Channel { get; set; }
-        internal static int? WaitTimeSeconds { get; set; }
+        internal static ulong? WaitTimeSeconds { get; set; }
         internal static string? OkCountEmoji { get; set; }
         internal static string? BadCountEmoji { get; set; }
     }
